+++ Tests/SentryTests/HybridSDK/SentryInternalScreenApiTests.swift
@@ -44,6 +44,7 @@ private class MockHub: Hub {
 
     func storeEnvelope(_ envelope: SentryEnvelope) {}
     func captureEnvelope(_ envelope: SentryEnvelope) {}
+    func captureNonTerminatingEnvelope(_ envelope: SentryEnvelope) {}
     func captureErrorEvent(event: Event) {}
     func setTrace(_ traceId: SentryId, spanId: SpanId) {}
     var currentOptions: Options? { options }